backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
sqlalchemy
pydantic
jinja2
python-multipart
openai
